Refactor components to named exports instead of defaults. Remove nested header folders - don't really need them
