+++ src/app/components/header/ContraLogo.tsx
@@ -2,7 +2,7 @@ import Image, { type ImageProps } from "next/image";
 import contraLogo from "./contra-logo-framed.png";
 import { type OptionalKeys } from "~/utilTypes";
 
-export default function ContraLogo({
+export function ContraLogo({
   height = 200,
   alt = "Round-framed image of 8 contra dance goers on a gold background.",
   className = "",
