--- src/app/components/header/mobileHeader/MobileHeader.tsx
@@ -1,11 +1,0 @@
-import MobileHeaderBadge from "./MobileHeaderBadge";
-import MobileNav from "./MobileNav";
-
-export default function MobileHeader() {
-  return (
-    <header className="relative flex flex-col desktop:hidden">
-      <MobileNav />
-      <MobileHeaderBadge />
-    </header>
-  );
-}
