--- src/app/components/header/mobileHeader/MobileNav.tsx
@@ -1,10 +1,0 @@
-import { MenuIcon } from "lucide-react";
-
-export default function MobileNav() {
-  return (
-    <MenuIcon
-      size={36}
-      className="absolute right-2 top-2 stroke-2 text-stone-800 text-opacity-60 hover:text-opacity-100 active:text-opacity-100"
-    />
-  );
-}
